// src/lib/api/generated/model/token.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Atuna API
 * A back-end for Atuna.
 * OpenAPI spec version: 1.0.0
 */
import type { TokenUserId } from "./tokenUserId";
import type { TokenRole } from "./tokenRole";

export interface Token {
  access_token: string;
  token_type: string;
  user_id?: TokenUserId;
  role?: TokenRole;
}