// src/lib/api/generated/model/clientRegister.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Atuna API
 * A back-end for Atuna.
 * OpenAPI spec version: 1.0.0
 */
import type { ClientRegisterFirstName } from "./clientRegisterFirstName";
import type { ClientRegisterLastName } from "./clientRegisterLastName";
import type { ClientRegisterPhone } from "./clientRegisterPhone";

export interface ClientRegister {
  email: string;
  first_name?: ClientRegisterFirstName;
  last_name?: ClientRegisterLastName;
  phone?: ClientRegisterPhone;
  password: string;
  role?: string;
}